Full and half screens switch windows when pressed twice
If for example you have a window taking up the right half of the screen 
and you press the hotkey to make the window take up the right half of 
the screen again, it will move over to the next window and take up the 
right half of the screen there. This only works if you have the dock and 
menubar hidden.