fix: next/previous window command inconsistency
The old method of getting the next and previous screen was inconsistent
about the order of the screens. Even if the screen layout remained
consistent, hammerspoon would reverse the next/previous order about
50% of the time after restarting the computer or reconnecting the
computer to external monitors. This new method is consistent and
more specific, moving screens left or right instead of next or previous.